Measure & COM table fix

+++ Modules/DeviceInfo/deviceInfo.h
@@ -12,8 +12,8 @@
 class DeviceInfo
 {
 private:
-	static uint16_t m_modbus1_value[MODBUS_SENS_COUNT];
-	static uint16_t m_1wire_value[MODBUS_SENS_COUNT];
+	static int16_t m_modbus1_value[MODBUS_SENS_COUNT];
+	static int16_t m_1wire_value[MODBUS_SENS_COUNT];
 
 public:
     typedef struct _info_t {
@@ -80,14 +80,14 @@ public:
     };
     struct modbus1_last_value
     {
-        static void set(uint64_t value, unsigned index = 0);
-        static uint64_t get(unsigned index = 0);
+        static void set(int16_t value, unsigned index = 0);
+        static int16_t get(unsigned index = 0);
         static unsigned index(unsigned index = 0);
     };
     struct _1wire_last_value
     {
-        static void set(uint64_t value, unsigned index = 0);
-        static uint64_t get(unsigned index = 0);
+        static void set(int16_t value, unsigned index = 0);
+        static int16_t get(unsigned index = 0);
         static unsigned index(unsigned index = 0);
     };
     struct _1wire_registrate
