+++ Modules/DeviceInfo/deviceinfo.cpp
@@ -4,8 +4,8 @@
 #include "clock.h"
 
 
-uint16_t DeviceInfo::m_modbus1_value[MODBUS_SENS_COUNT] = {};
-uint16_t DeviceInfo::m_1wire_value[MODBUS_SENS_COUNT] = {};
+int16_t DeviceInfo::m_modbus1_value[MODBUS_SENS_COUNT] = {};
+int16_t DeviceInfo::m_1wire_value[MODBUS_SENS_COUNT] = {};
 DeviceInfo::info_t DeviceInfo::info = {};
 
 
@@ -105,12 +105,12 @@ uint64_t DeviceInfo::need_registrate_1wire::get(unsigned)
 }
 
 
-void DeviceInfo::modbus1_last_value::set(uint64_t value, unsigned index)
+void DeviceInfo::modbus1_last_value::set(int16_t value, unsigned index)
 {
 	m_modbus1_value[index] = value;
 }
 
-uint64_t DeviceInfo::modbus1_last_value::get(unsigned index)
+int16_t DeviceInfo::modbus1_last_value::get(unsigned index)
 {
     return m_modbus1_value[index];
 }
@@ -121,12 +121,12 @@ unsigned DeviceInfo::modbus1_last_value::index(unsigned index)
 }
 
 
-void DeviceInfo::_1wire_last_value::set(uint64_t value, unsigned index)
+void DeviceInfo::_1wire_last_value::set(int16_t value, unsigned index)
 {
 	m_1wire_value[index] = value;
 }
 
-uint64_t DeviceInfo::_1wire_last_value::get(unsigned index)
+int16_t DeviceInfo::_1wire_last_value::get(unsigned index)
 {
     return m_1wire_value[index];
 }
